requirements.txt
Flask
openai-whisper
pydub
torch
